pyproject.toml
[project]
name = "crude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt[duckdb]>=1.1.0",
    "duckdb>=1.1.1",
    "httpx>=0.27.2",
    "joblib>=1.4.2",
    "numpy>=2.1.1",
    "plotly>=5.24.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "streamlit>=1.38.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "watchdog>=4.0.2",
]
